requirements.txt
python-telegram-bot[job-queue]==20.0
deep-translator
requests